check_multiple.py: overhaul to support python-3.x.
Tested with python-3.6, but I've tried to make choices backwards-
compatible to python-3.5. Specifically, we use the subprocess.run()
method that only exists in python-3.5, but avoid the "capture_output"
and "text" keywords to that method that only exist in python-3.7.

A comprehensive list of changes would be too long to include here. A
large number of minor refactorings were done in the process of the
large refactorings necessary to adopt the python-3.x APIs.