Support WARNING as well, and return the best/worst overall status.
Instead of forcing a result of either OK or CRITICAL, it may sometimes
be desireable to return WARNING. For example, in mode "one", if all
checks return WARNING, you probably want an overall result of WARNING
as well, because that's the "best" status you've seen.

This commit adds support for reporting and returning warnings, and
updates the test suite to check all 3^3 combinations (in both modes)
of individual statuses.