check_multiple.py: switch from one/all modes to worst/best.
Now that we have the ability to handle WARNING results, the modes
"one" and "all" are misnomers. Instead what we're really returning is
the best or worst overall result. In the past, with only two
possibilities, we had best == one and worst == all. But now we want
probably want to return WARNING and not CRITICAL if every individual
check returns WARNING. The names "best" and "worst" make it easier
to tell the program what you want.