Refactor into separate library and executable.
This is a lot of work to solve a stupid problem: when deploying the
executable, there's no good way to drop the ".py" extension on-the-fly
to wind up with a command named "check_multiple" as the README suggests.
Of course, I can rename the script from "check_multiple.py" to simply
"check_multiple", but that breaks the test suite, which insists on
being able to import the module containing the tests (which it can't
do if the "module" is missing its .py extension).

To ensure that the command can be called with "check_multiple", and to
keep the test suite working, this commit splits the executable and
library into two files. The executable now lives at bin/check_multiple,
and is extensionless. The library, on the other hand, lives in its own
package, lib/check_multiple. The check_multiple.py module within that
package can be imported by python, and can therefore be tested.

Having the two files separate is a bit annoying, because now you can't
just run the thing during development, but there's an easy PYTHONPATH
hack to make that work. I've added it to the README.